OplatiPay_v2/requirements.txt
python-telegram-bot==20.3
openpyxl
requests